% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluate_polynomial.R
\name{eval.pol}
\alias{eval.pol}
\title{Evaluate a polynomial in matrix form}
\usage{
eval.pol(pol, val)
}
\arguments{
\item{pol}{The objective polynomial represented in matrix form.}

\item{val}{The values of variables, from x_1 to x_n.}
}
\value{
Numeric.
}
\description{
Evaluate a polynomial in matrix form
}
\examples{
polynomial <- pol(5,1:36) ##Create a polynomial with 5 variables. The
coefficients go from 1 to 36 for the 36 terms.
}

