% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/final_chart.R
\name{final}
\alias{final}
\title{Calculate the final result of the three different methods with different numbers
of connections.}
\usage{
final(n = 20, expen = 50, times = 1000, cost = 0)
}
\arguments{
\item{n}{The number of variables (inputs) in the polynomial. Default: n=20}

\item{expen}{The total amount of expenditure. Default: expen=50}

\item{times}{The number of times that you want to repeat for each of the functions
so that the result will be more precise. Default: times=1000}

\item{cost}{The cost at each point. Default: cost=0}
}
\value{
data.frame The result of calculations. See details.
}
\description{
Calculate the final result of the three different methods with different numbers
of connections.
}
\details{
Such function generates a whole chart of results. Each time a random set
of parameters will be generated from [-1, 1] and a random set of inputs will be
chosen.
The local maximum for each of the three strategies wil be calculated, and such
maximum will be divided into five cases. The number of connections increases from
1 to 5. Each row of the chart represents the outcome of a strategy, and each column
represents the outcome of a specific number of connections.
}
\examples{
final()
final(n=5, expen=10, times=100, cost=0.1)
}

