% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_polynomial.R
\name{pol}
\alias{pol}
\title{Create a polynomial in matrix form}
\usage{
pol(n, c)
}
\arguments{
\item{n}{The number of variables.}

\item{c}{A string of the coefficients of the polynomial. The length of c should
equal to (n^2+3n+2)/2}
}
\value{
A matrix. Column and row names represent variables, and the value of
each individual slot represents the coefficient of the multiple of its
corresponding row and column names.

The first variable x_0 is set to be 1.
}
\description{
Create a polynomial in matrix form
}
\examples{
pol(5, 1:21)
}

