% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate_points.R
\name{random.input}
\alias{random.input}
\title{Create a random string of input}
\usage{
random.input(n, expen)
}
\arguments{
\item{n}{The number of different inputs.}

\item{expen}{The total amount of expenditures.}
}
\value{
A string of numbers that represent different amount of inputs.
}
\description{
Create a random string of input
}
\examples{
random.input (inp = 20, expen = 50)  #The total expenditure of 50 is assigned
to 20 different inputs.
}

