% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/repeat_steps.R
\name{repeat.climb}
\alias{repeat.climb}
\title{Calcilate the mean and standard deviation of the results from a specific case}
\usage{
repeat.climb(n, expen, times, connection, kind = c("steep", "median",
  "least"), cost = 0)
}
\arguments{
\item{n}{The number of variables (inputs) in the polynomial.}

\item{expen}{The total amount of expenditure.}

\item{times}{The number of times that you want to repeat for each of the functions
so that the result will be more precise.}

\item{connection}{The number of connections.}

\item{kind}{One of the three functions that you would like to choose.}

\item{cost}{The cost at each point.}
}
\value{
a string of the form "aaaa-bbb".
}
\description{
Calcilate the mean and standard deviation of the results from a specific case
}
\details{
In the return, "aaaa" represents the mean value of all the local maximums
and "bbb" represents the standard deviation of those maximums.
}
\examples{
repeat.climb(n=5, expen=10, times=50, connection=2, kind="steep")
}

