% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/what.is.a.step.forward.R
\name{step.choice}
\alias{step.choice}
\title{Figure out all the results when you take a step forward.}
\usage{
step.choice(pol, val, connection)
}
\arguments{
\item{pol}{The objective polynomial represented in matrix form.}

\item{val}{The values of variables, from x_1 to x_n.}

\item{connection}{The maximum distance you can transfer a unit of expenditure from
one input into another.}
}
\value{
A matrix. The first column represents the result of a transfer, the
second column represents the input you are reducing, and the third represents
the input you are increasing.
}
\description{
Figure out all the results when you take a step forward.
}
\examples{
step.choice(pol = pol(5,1:21), val = (1:5), connection = 2)
}

