% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/take.many.steps.R
\name{step.final.steep}
\alias{step.final.steep}
\title{Calculate the highest point when climbing along the steepest direction}
\usage{
step.final.steep(pol, val, connection, cost = 0)
}
\arguments{
\item{pol}{The objective polynomial represented in matrix form.}

\item{val}{The values of variables, from x_1 to x_n.}

\item{connection}{The maximum distance you can transfer a unit of expenditure from
one input into another.}

\item{cost}{= 0 The money you would spend on each points.}
}
\value{
Numeric. The highest profit you can ever reach when you choose a
specific point and climb along the steepest direction.
}
\description{
Calculate the highest point when climbing along the steepest direction
}
\details{
There are three functions. When you apply the function step.final.steep,
everytime when you find a list of available choices, such function will choose the
one that increases the value of the polynomial most. The function step.final.med
will choose the median among the whole least, and step.final.least will choose the
least one.
}
\examples{
step.final.steep(pol = pol(5,1:21), val = (1:5), connection = 2, cost = 0,1)
}

